#43: Move DataImit to gost89
